chore: 묵찌빠관련 함수 MukChiBbaGame파일추가 후 이동
